requirements.txt
boto3==1.38.45 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.38.45 ; python_version >= "3.9" and python_version < "4.0"
certifi==2025.6.15 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.10 ; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
nuvla-api==4.2.3 ; python_version >= "3.9" and python_version < "4.0"
paho-mqtt==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.9" and python_version < "4.0"
s3transfer==0.13.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.20 ; python_version >= "3.9" and python_version < "3.10"
urllib3==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
jupyter
numpy
pillow
protobuf
requests
torch==2.4.1
dontenv
pathlib
